I'm thinking about moving the "resource" feature into a new plugin.
This is just WIP and I'm not sure if I'm going to do it in the future.
This feature is really nice, but is something I have not used in two
years and a half. The idea behing it was to be able to manage stuff
which is non-active_record, for example Git, Memecahced ...